requirements.txt
asgiref==3.8.1
asttokens==3.0.0
bleach==6.2.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_ag_grid==31.3.0
decorator==5.1.1
Django==5.1.4
django-cors-headers==4.6.0
django-fernet-encrypted-fields==0.1.3
django-picklefield==3.2
django-plotly-dash==2.4.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
dpd-static-support==0.0.5
dpd_components==0.1.0
executing==2.1.0
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
ipython==8.31.0
itsdangerous==2.2.0
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.5
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.1
openapi-codec==1.3.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
plotly==5.24.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
retrying==1.3.4
setuptools==75.6.0
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.0.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
whitenoise==6.8.2
zipp==3.21.0
